+++ src/vue/pages/Business.vue
@@ -14,7 +14,7 @@
           class="business__login-link"
           :to="vueRoutes.login"
         >
-          <img class="business__logo" src="/static/conto-logo.png">
+          <img class="business__logo" src="/static/logo-black.svg">
           <img class="business__small-logo" src="/static/favicon-conto.png">
         </router-link>
       </div>